voiceAuth/requirements.txt
fastapi
uvicorn
resemblyzer
pydub
numpy<2
PyJWT
python-multipart
torch==2.2.2+cpu
torchaudio==2.2.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
